+++ apps/backend-starter/src/app.module.ts
@@ -1,11 +1,10 @@
+import { ModelsModule } from '@generated/nestjs-models';
 import { Module } from '@nestjs/common';
 import { APP_FILTER } from '@nestjs/core';
 import { LoggerModule, PrismaExceptionFilter } from '@tractr/nestjs-core';
 import { DatabaseModule } from '@tractr/nestjs-database';
 
 import { AppController } from './controllers';
-
-import { ModelsModule } from '@generated/nestjs-models';
 
 @Module({
   imports: [LoggerModule, DatabaseModule.register(), ModelsModule.register()],