+++ packages/angular-tools/projects/angular-tools/src/lib/utils/unsubscriber.ts
@@ -1,0 +1,12 @@
+import { Component, OnDestroy } from '@angular/core';
+import { Subject } from 'rxjs';
+
+@Component({ template: '' })
+export abstract class Unsubscriber implements OnDestroy {
+  protected unsubscribe$: Subject<void> = new Subject<void>();
+
+  ngOnDestroy(): void {
+    this.unsubscribe$.next();
+    this.unsubscribe$.complete();
+  }
+}