+++ packages/nestjs-authentication/src/authentication/interfaces/authentication-module.interface.ts
@@ -1,9 +1,8 @@
 import { JwtModuleOptions } from '@nestjs/jwt';
 import { IAuthModuleOptions } from '@nestjs/passport';
+import { User } from '@prisma/client';
 import { StrategyOptions } from 'passport-jwt';
 import { IStrategyOptionsWithRequest } from 'passport-local';
-
-import { User } from '@prisma/client';
 
 export interface AuthenticationOptions extends Record<string, unknown> {
   login: {