+++ packages/nestjs-core/CHANGELOG.md
@@ -3,6 +3,13 @@
 All notable changes to this project will be documented in this file. See
 [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# @tractr/nestjs-core [1.5.0](https://github.com/tractr/stack/compare/@tractr/nestjs-core@1.4.1...@tractr/nestjs-core@1.5.0) (2021-04-12)
+
+
+### Features
+
+* update README of packages to publish new versions ([#77](https://github.com/tractr/stack/issues/77)) ([e6bf415](https://github.com/tractr/stack/commit/e6bf415af3fe5588c15577f047a6262f81c1564f))
+
 ## @tractr/nestjs-core [1.4.1](https://github.com/tractr/stack/compare/@tractr/nestjs-core@1.4.0...@tractr/nestjs-core@1.4.1) (2021-04-05)